K8s Updates: Include certificate rotation
Certificates and relatives are now handled by vault.
Previously, we could be sure that during Kubernetes Updates,
certificates get renewed. As they are now no longer managed by kubeadm,
we need to execute the k8s-master and k8s-worker roles for them to get
renewed. That way we can still be sure that certificates get renewed if
they are close to expire during Kubernetes updates.