Fix k8s v1.25 Tigera installation version
During an K8s upgrade play, we have to overwrite "k8s_version",
otherwise the calico version of Kubernetes v1.24 will be used as we're
mapping the calico version to the K8s version in "k8s_config".
However, we want to install the calico version mapped to v1.25 right
away.