k8s-master: fix name of front proxy client certificate
As it turns out, Kubernetes has strong opinions about the commonName in
that certificate, at least in some applications. In particular, the
desec-webhook refused to work because it also checked the commonName to
equal front-proxy-client.

This needs manual intervention in existing vaultified clusters, but as
those don't really exist yet, we do not handle a migration path here.