rook: Workaround custom ServiceMonitor labeling
Currently it is not possibly to pass custom labels for ServiceMonitors
via the helm chart [1].
However, we're supporting the feature to limit to let Prometheus scrape
only specifically labeled ones. In clusters using this feature, the
rook-ceph-mgr ServiceMonitor won't get picked up.

To workaround this, we're labeling all rook components with the
common_labels which will result in properly labeled ServiceMonitors.

[1] https://github.com/rook/rook/issues/9618