containerd: use systemd cgroups also with nvidia plugin
The SystemdCgroup option is scoped to the runtime instance. As we create
a separate runtime instance for nvidia, we have to make sure to also set
that option there. Otherwise, containerd will not use systemd unified
cgroup magic to start the pods which will cause SandboxChanged events
and pod kills all over.

These are notoriously hard to debug and the only evidence, aside from
the SandboxChanged event in the pod and ominous "Killing container with
a grace period" log messages in kubelet there is nothing to go by unless
you turn up verbosity to 5 and/or add some context to your grep (then
you'll also find "The cgroup has some missing controllers", but
without direct mention of the pod, making it tricky to find).

You may then also find messages like 'Factory "containerd" was unable to
handle container' or mention that a specific pod is not handled by
systemd, which should put you on edge already.

Anyway, one line fix for a major problem once more. What finally pointed
me to the solution was https://gjhenrique.com/cgroups-k8s/.