Enter venv from .envrc
This commits adds a function to the .envrc to automatically enter an
isolated Python environment upon entering the cluster repository,
automatically installing and upgrading dependency versions if necessary.
It adds a hook to watch the dependency file to ensure that the current
venv is always up-to-date with the checked-out revision of yk8s.