vault: Import etcd backup configuration
This already handled in the import.sh script for existing clusters,
but not for freshly created clusters or clusters with intermediate CAs.
For fresh clusters, one would have to manually put the etcd backup
credentials into vault, which is not obviously and complicated.
Allowing to put the credentials into a yaml file which is then
automatically handled and removed after the vault import is more
convenient.