k8s: Bump mapped calico version to 3.26.1
By default we're using ED25519 for the Root CAs after feature/vault.
Calico workload can't handle these (rel [1]) for versions >3.25 <3.26.1
Although the fix is going to be backported to 3.25, we're directly
bumping to the latest calico version for Kubernetes v1.25 so that we
do not have to wait for that.

This will trigger calico upgrades on existing clusters with K8s v1.25
on the next stage3 run if they don't configured a custom calico version.
This is harmless though.