Add flake.nix
Nix is a declarative package manager which powers NixOS but can also be
installed as an additional package manager on any other distribution.
This commit adds a flake.nix which references all necessary
dependencies which are locked to specific versions.

Together with poetry (which is also installed by the flake), this
allows for 100% reproducible and (should you chose to) pure
environments shared between all users.