Drop legacy monitoring_v1 role
monitoring_v1 deployed the prometheus stack via jsonnet magic.
That was the standard installation method back in the old days.
Now we got helm which is doing a much better job.
monitoring_v1 is marked legacy for a long time and not used in the wild
anymore as far as I know.