GPU: Do not uninstall nvidia device plugin of no GPUs targeted
If no gpu nodes are targeted in the current play,
k8s_nvidia_device_plugin_install will became false and the
nvidia-device-plugin will get uninstalled.
However, one may limits e.g. the targets such that no GPU node is
targeted in the current run, but is still part of the Kubernetes cluster
which would cause havoc.

We don't want the nvidia-device-plugin to get uninstalled in that case.
We want it to remain until k8s_is_gpu_cluster is false.

This is in especially the case during the Kubernetes upgrade to v1.25
where we want to rollout restart the nvidia-device-plugin if the
migration to the Tigera operator has happened (which it is by default).
As facts are available on play level only, we would need to target the
play at masters:workers, which would require to implement some
workarounds as we e.g. don't want calicoctl on worker nodes.