rook_v2: Don't verify each pod during health checks
Important is, that the Ceph cluster is healthy. Checking each Pod is
error-prone, as their may be temporary Pods triggered by Jobs or similar
which are not expected to be Running all the time and if we check the
state of these Pods in an unfortunate moment, ansible may fatally fail.