Remove the python-openstackclient requirement for baremetal deployment
**DO NOT MERGE THIS COMMIT TO THE MAINLINE**

This requirements breaks the baremetal deployment process on Ubuntu 22.04 machines.

The reason is that it depends on the no longer maintained netifaces
module. As it is no longer maintained and the last release is from May
2021, there are no wheels for python 3.10 – however the system python on
22.04 is 3.10.

In consequence the deploy process tries to build the package from
source, which it can't because we don't have the build dependencies
installed.

Not depending on python-openstackclient is no problem in a baremetal
setting, the openstackclient is only used for cleanup for clusters atop
OpenStack.

The issue will be fixed by the core split, as the parts of mk8s used on
baremetal deployments will no longer depend on the openstack and
terraform parts.