Revert "Allow inplace Kubernetes upgrades"
This reverts commit 048ad2340411bd81c2ea61dbf16f016e9e2b31bd.

Inplace updates do harm to RWO PV attachments, at least for the k8s
update from 1.24 to 1.25. We should not diverge our update procedure
from the officially supported instructions.

Each attachment of a PVC to a Node is tracked by a VolumeAttachment
object. That object carries state information such as the current
attachment state, backend-specific information (e.g. volume
IDs, server IDs, device names) and which node and volume it is about.

The inplace update to 1.25 messes up the state so that RWO PV can't be
attached to other nodes anymore and one needs to manually clean up.