rook_v2: Fix rescuing health verification loop
Block tasks can only be directly be rescued, if a task in an included
tasks file fails, it won't be rescued. Therefore, the cluster health
verification tasks are put together in one block in one file.

Furthermore, the retry count is increased to 50 and a short pause of 5s
has been added in-between checks.