rook_v2: Rework upgrade checks
The debug module + failed_when tasks led to fatal Ansible errors in the
past which then completely stopped the playbook.

Instead, the assert-module is now used to verify if the ceph cluster has
been updated, yet. The assert-module may print the whole deployment
manifest, which may be a few hundred line sof logs, therefore, no_log is
set.

As we don't have a rook-to-ceph-version map anymore and allow to set a
custom ceph version, we can't check if the ceph version has been rolled
out since we do not know the reference version. Instead we could check
if all deployments use the same Ceph version, which is a future task.