Write rook variables also to stage5 inventory
I don't know how this has worked out previously, but after dropping the
outdated monitoring_v1 role, the creation of the ServiceMonitor for the
rook ceph manager fails.
However, previously only some defaults values were use, although it is
not entirely clear to me where they came from.
To fix and create a proper ServiceMonitor even for customized rook
namespaces and clusternames, we have to write the rook variables also to
the inventory of stage5.

A more sophisticated fix is out of scope as we're going to rethink the
stages anyway in the so called core-split.