CI: Adjust for yaook/k8s-ci image with poetry
As Debian 12 does not allow to install pip packages outside of a virtual
environment anymore, we came to the conclusion that we should take that
chance and setup poetry in the k8s-ci image directly [1].
Therefore, there is no need to install poetry in each job anymore.

We want to do a "poetry install --sync" nonetheless, i.e. for MRs
tinkering with the dependency declarations.

[1] https://gitlab.com/yaook/images/k8s-ci/-/merge_requests/54