Make vault.sh work with rootless docker/podman
vault.sh runs the container as the current user so the owners of
./vault/* line up. This does not work in rootless setups, however,
where uids inside the container are mapped to a subuid range.

This commit adds a new user-specific environment variable
VAULT_IN_DOCKER_USE_ROOTLESS which if set to true will cause vault.sh
to spawn a chown job prior to vault in order to fix the file
permissions.