Restart nvdp Pods on K8s upgrades
The nvdp marks a GPU as unhealthy on systemctl reloads and a following
kubelet restart. This is a bug of the nvdp. The Pod does not fail and
the GPU is not usable anymore. This happens during Kubernetes upgrades,
therefore we have to restart each nvdp on k8s nodes with GPUs before
we uncordon them. This ensures that the impact on workload is kept minimal.