snapshot-controller: Lay the base for future versions
This commit lays the basis to support future versions of the volume
snapshot controller. The version is mapped to the configured Kubernetes
version in the k8s-config role and updated also during Kubernetes upgrade
plays. Therefore, we should keep track of new versions available when
adding support for new Kubernetes versions.