Add support for Kubernetes v1.26
This commit adds support for Kubernetes v1.26. Versions of components
mapped to the Kubernetes role have been updated to be up-to-date.

We're enforcing the migration to the calico Tigera operator on upgrades
to Kubernetes v1.26. This step has been announced in [1]. This is a
necessary step to get rid of the manifest-based installation method at
some point in time, respectively when dropping support for Kubernetes v1.25.

When adding support for the OpenStack CCM and cinder-csi-plugin for
Kubernetes v1.26, I decided against using symlinks by purpose although
the templates did not change. The reason for that is, that we can then
easily drop support for Kubernetes v1.25 at some point in time without
having to tinker with the symlinks. As the templates don't get touched
at all normally, we do not have to bother about code duplication with
inconsistencies that much.

[1] c1042793a2a9541fb9de31cfdc18ef97717de829