rook: Fix common monitoring label iteration
The default needs to be a dict, not an array.

The previous condition led to errors when iterating over the list of
common monitoring labels if the default is used due to items not being
available for arrays.

Although this is fixed by adjusting the default, checking if the dict
is empty instead of checking if it is defined is a better check.