ipsec: include passwordstore role only if enabled
The ipsec role hasn't been fully migrated to vault yet and still depends
on the passwordstore role. If ipsec is not used, initializing a password
store is not necessary. However, as an ansible dependency, it was still
run and thus failed if passwordstore hadn't been configured. This commit
adds the role via `include_role` instead of as a dependency.