Add nodeSelector to Tigera operator
As the Tigera operator is a system component we want to ensure it is
spawned on the control plane so that it can not be evicted by worker
node pressure or similar. This is in especially a necessity as the
helm chart currently does not allow one to configure a PriorityClass.