Fix Thanos Terraform-managed bucket setup
If one defines a cluster name in the terraform section and uses Thanos
with a managed bucket (manage_thanos_bucket=true which is default),
the previous templating created incorrect secrets for Thanos.
Terraform will create a bucket in the form of
"{{ cluster_name }}-monitoring-thanos-data".
As the cluster_name previously has not been populated into the Ansible
stages, the inventory updated had to be adjusted.
During or post core-split we should rework our config handling.