CI: Fix ansible-lint invocation
The syntax and command intepretaion changed and if one explicitly passes
a playbook/role only (sic!) that file will be linted and ansible-lint
will not (!) lint roles/files mentioned in the passed lintable.
To invoke a hollistic linting process with nested invocation, one must
not pass files (lintables) so that the auto-dection mode is triggered.

The alternative would be to link to every playbook and role in the
repository which is obviously not a real alternative.

This has been broken since at least jumping to ansible-lint > 6.15.0.

https://ansible.readthedocs.io/projects/lint/usage/