ansible-lint: Add further rules to skip list
var-naming[no-role-prefix]
Would make it mandatory that each variable has its role name as prefix.
Renaming all the variables is not feasible and would probably take a
lot of time to get the LCM back into a working state. Therefore, skip it.

fqcn-builtins
Do not make it mandatory to use the full fqcn for builtin modules.
During discussions we agreed that this does not improve readability.
It couldn't be explicitly ignored for builtins only before though

yaml[document-end]
The enabled auto-dection mode now scans more files. It would be
cumbersome to edit e.g. CRD-files and similar. We may re-enable that
in the future, though.