fluxcd: Use the community helm chart by default
The old installation method is hard to maintain. This commit introduces
a new role which migrates and installs flux via the community helm chart
by default therefore. Helm charts are used for nearly every other
service we're supporting. Maintaining manifests on our own is not
feasible. Migrating to the community helm chart will make it easier for
us to keep up-to-date.

For the migration, we're adding labels and annotations to existing
resources so that they will be adopted.
The migration is guarded with a simple, but hopefully sufficient check.

The PodMonitor needs additional labels, otherwise it won't be picked up.