Ignore release note check for renovate MRs
It is not trivial to convince the renovate bot to create a release note
file along with a merge request. One issue is the typscript regular
expression [1] which does not like to work with '$' (even with escaping),
but furthermore the command templating may also fail with a nested
command. We need to generate a file with a unique name though.

To stop working on a problem which may not even is one, we decided to
let the release-note-file-check be skipped for merge requests which
look as they come from the renovate bot. Therefore, renovate merge
requests will not be listed in the release notes for now.

We can not skip explicitly for the renovate bot user ID as on merge
trains this will be the one of the trigger.

https://github.com/renovatebot/renovate/blob/d847715d2c7b3f7efdb10937acbc7e09e143c5f5/lib/workers/repository/update/branch/execute-post-upgrade-commands.ts#L70