Add option to configure custom DNS nameserver for cluster subnet
We're introducing the option to set custom DNS nameservers for the
IPv4 cluster network subnet on OpenStack. This is useful if one does not
want to use the default ones. We're introducing this to IPv4 only for now
as the DualStack-support still is broken.
However, we do not validate if the list is in a correct format. This will
be done by Terraform itself.

Closes: #594