kubeadm: Use kubeadm.k8s.io/v1beta3 configuration
We're reworking the initialization and join configuration as there is
a new configuration format available. This allows us to ignore pre-flight
errors via the configuration files instead of having to pass them as
CLI flags to the kubeadm invocation.
This is also a hard prerequisite for the Kubernetes v1.27 support,
as Kubernetes v1.27 won't support kubeadm.k8s.io/v1beta2 anymore.

https://kubernetes.io/docs/reference/config-api/kubeadm-config.v1beta3/#kubeadm-k8s-io-v1beta3-JoinConfiguration