prometheus: Allow the usage of persistent storage
By default, Prometheus is configured to use empty-dir volumes. This
causes losses of metrics-data if the Pod gets terminated i.e. during
maintenance work like Kubernetes upgrades or upgrades of the Prometheus
stack itself. If Thanos is configured along Prometheus as long term
storage, not all metrics get lost but only these which were not part of
the last retention. This can be up to two hours of metrics.
Due to compliance reasons it is often required to have gapless metrics
available.

This commit allows the specification of a storage class and the
respective volume resource request size for persistent storage for
Prometheus. It is disabled by default though, as the availability of
StorageClasses varies depending on the environment and therefore we
cannot set a default. A size of 50Gi is the default value.