core-split: Split k8s-api-frontend role
This role tinkers with HAProxy and nftables. HAProxy is used in general
for a high-available Kubernetes Control-Plane whereas nftables is only
deployed on mk8s-Cluster running on OpenStack and with Gateways in front
which means the haproxy-role is moved to k8s-core whereas the nftables-
role is moved to k8s-supplements.
Therefore, we must split this role. Having a separate role for these
tasks does not make sense anyway.
In addition, the haproxy-role has been a bit reworked for better
deployment.