core-split: Merge k8s-bgp and bird role and move to k8s-supplements
There is no need to have separate roles for these. Previously, it was
a necessity due to the stage separation and because the bird role had
been executed in stage2 but the k8s-bgp role requires has been executed
in stage3 (and needs variables available only there).

Now we can bundle these roles together.