destroy: only delete OpenStack objects belonging to the project
This filter should normally not be necessary, because the credentials
are normally scoped to a project. However, at least Neutron will list
*all* resources from *all* projects if the *user* has sufficient
privileges for that, no matter the project scoping of their token.

This has caused absolute chaos with C&H's Kubernetes customers when a
DevOps person accidentally had the admin role in their project and they
deleted their cluster.