Use OpenTofu as in-place replacement for Terraform
As HashiCorp changed the license of Terraform from MPL to BUSL, we're
switching to the community-driven in-place replacement fork OpenTofu
to avoid legal risks.

This commit replaces all mentions of Terraform in the documentation with
OpenTofu, replaces terraform with opentofu in the dependencies and adds
a wrapper script to ensure backwards compatibility.

Addresses: #601

https://opentofu.org/