Rework package installation for pkgs-k8s-io
We're switching over to use the Kubernetes community-owned software
repositories as the Google-hosted ones got frozen.
Additional work had to be done as there now exists one repository for
each minor version. Furthermore, the revision part (-00) is now auto-
generated and has a new format.
The repository configuration has been put into an own tasks file so that
it can be triggered during Kubernetes upgrades therefore (as we now have
to configure the repository when updating to a new minor version).

It is important to install all packages along as e.g. kubeadm depends on
kubectl and if installed separately, a newer version then desired may
gets installed which would require to downgrade. That's obviously not
desired.

The deprecated tasks which allowed the installation via other package
managers than apt have been dropped along this way as we're currently
supporting Debian-based operating systems only.

Closes: #587

https://kubernetes.io/blog/2023/08/15/pkgs-k8s-io-introduction/