ch-k8s-lbaas: change pull policy to IfNotPresent
Always has downsides:

- It is slower.
- It may break if the registry is unavailable.

It also has upsides, namely that you can replace an image in place and
still... wait, that's a downside, because it's a thing which shouldn't
be done anyway unless you are in a development situation in which case
you can also just kubectl edit the deployment to use Always manually. So
no upsides. Let's go with IfNotPresent.