Set kubeconfig for all action scripts
As it is a prerequisite for the LCM that the orchestrator node can
interact with the Kubernetes API during rollouts. It is put into lib.sh
because that file is sourced in all other action scripts.

In pre-core-split times, we did that only for stage{4,5} as these did
interact with the Kubernetes API. However, the new abstraction format
does not allow such differentiation. Both, k8s-core and k8s-supplements
require that the orchestrator can interact with the Kubernetes API.