Skip vault-onboarded tasks if already checked
This role is triggered multiple times during an apply-all rollout as it
is part of the connect-to-nodes.yaml playbook. As the connection
establishment is a prerequisite for all other playbooks and the apply-all
playbook triggers them all, the vault-onboarded tasks get executed
multiple times during a apply-all rollout which is time-consuming.

The connect-to-nodes.yaml playbook has to be included by all other
playbooks as they must be functional on their own. The fact that this
led to multiple invocations during an apply-all is a downside of the
modularization of the LCM.