Set *_enabled variable for services by default
As we're merging the config template into the actual configuration,
this ensures that these variables are always set and a playbook does not
fail because these variables are undefined.

As the playbook structure has been reworked with the core-split,
i.e. the etcd-backup role failed if etcd-backups were not explicitly
disabled. This was the trigger to set these variables for everything
in the template.