Move IPSec EAP PSK to vault
We want to drop the passwordstore very soon. The IPSec pre-shared key
is one of the few things which still need to be moved to vault.
The migration is kept relatively simple, although we can't avoid to
update the gateway and orchestrator policies. I decided to give the
orchestrator role only read privileges, because the value does not have
to get updated frequently but usually just once (and if rotated).