Auto upgrade terraform plugins while running terraform init
The gitlab-managed terraform backup does not work, if there is an
inconsistense in the versions of terraform plugins: while rerunning
`init` the new modules are installed but the old ones are not changed.

With the flag `-upgrade` terraform will upgrade all previously-selected
plugins to the newest version that complies with the configuration's
version constraints. This will cause Terraform to ignore any selections
recorded in the dependency lock file, and to take the newest available
version matching the configured version constraints.