smoke-tests: Speed up calico pings
This test takes very long to succeed. Therefore, the amount of Pods is
decreased, fewer packets are sent and we decrease the waiting time.
Although this smoke tests was implemented to check that all nodes can
each all other nodes, this test can be super time consuming in big
clusters and therefore the replica count of the deployment already has
been decreased. To speed things further up, the above actions are taken.
Inter-connectivity problems on the node layer will also be noticed by
other problems and in general one does not need to run this test on all
nodes. We're doing this for the sake of speeding up the smoke test suit
in whole and therefore also the CI pipelines.