smoke-tests: Rework explicit testing nodes
If nodes are explicitly listed as testing nodes, the service tests
should run on these only. At least two testing nodes must be defined
as e.g. the blocked network test won't work if scheduled on the same
node as well as e.g. the shared-fs tests don't make sense if executed
on the same node.

As it was unnecessarily complicated to define these nodes, the config
has been reworked such that the testing nodes can be passed as an array.
As appropriate label we can then just use the "kubernetes.io/hostname"
label.

For the check-shared-fs role we're using the first and last node of the
k8s_nodes group if no testing nodes are explicitly defined to remain
compatibility with a single node setup (although the test does not really
make sense on such a setup).