Pin to K8s v1.24 compatible version
As we're having clusters in the wild still running with Kubernetes v1.24
and we want to save us from the burden to do a dont-install-flux-with-
v1.24-dance in the CI, we're pinning to a Kubernetes v1.24 compatible
version by default now. One can easily overwrite this version if a newer
version is desired.

Furthermore, the yaook/operator part still needs Kubernetes v1.24 which
prevents us additionally from simply dropping support for it.