Terraform: Pull backup file of Gitlab state
A downside of the Gitlab backend is that a (malicious) user with
sufficient privileges may deletes the Terraform state from Gitlab (by
accident) which would cause havoc. Therefore, functionality has been
added which creates a local copy of the latest available state after
each Terraform action run.
This file is gitignored though to not having to track its changes in
git and to avoid confusion (as one configured HTTP as backend).
Operators may take additional actions to ensure availability in a
disaster scenario.