Fix nftables forwarding for multiple Wireguard endpoints
If multiple Wireguard endpoints are configured, we need to allow
forwarding for each of their interfaces. Otherwise, one can talk to the
Kubernetes API, but has no access to the internal network and therefore
can't ping or SSH nodes. A set is declared for a better overview.
This has been overlooked in !795.