Explicitly wait for Pod restarts
We have to wait until Pods are restarted. The implemented approach has
been chosen as we need multiple tasks and run them in a loop.
The status.containerStatuses[].state.running.startedAt key holds a sting
containing the latest container restart. As we're restarting the
containers (via the handlers), this is the value we have to check. As
we're delegating to localhost we have to explicitly set utc=true for the
now() function. The other values are chosen relatively random. Checking
if it restarted in the last 3m sound as a good starting point. We may
tweak the exact values though.