Add sonobuoy to CI
Sonobuoy is a diagnostic tool for K8s, that includes
Kubernetes conformance tests. We use Sonobuoy in
certified conformance mode.

As the test lasts for hours, it can not be run on every commit. So we'll
run the job on scheduled pipelines and when we create a new release.

As we use tainted nodes in our testcluster and sonobuoy can not handle
that [1] we need to untaint our gpu and storage nodes. In a future MR we'll build a
smaller cluster with untaintet nodes for faster testing.

As sonobuoy itself does not exit with an error if a plugin has failed
tests, we examine the results and fail if one plugin returned failed as
status.

[1] https://sonobuoy.io/docs/v0.57.1/faq/#we-have-some-nodes-with-custom-taints-in-our-cluster-and-the-tests-wont-start-how-can-i-run-the-tests