Move admin kubeconfig pull
Copy the admin kubeconfig directly after generating it. This is needed
if it gets renewed, but not by triggering the whole k8s-master role
but just invoking "obtain-kubeconfigs.yaml". If that's the case, a new
admin kubeconfig may gets generated but not copied to the localhost.
This happens during a CA rotation where we do not trigger the whole
k8s-master role but only parts and renew kubeconfigs.