Move CA ConfigMap creation
During a CA rotation we do not trigger the whole k8s-master role, but do
invoke just parts of it. During a CA rotation the CAs change and the
ConfigMap has to be updated. Therefore, we must move this task to
obtain-cas.yaml to keep the ConfigMap up-to-date without the necessity
to trigger the whole k8s-master role.