Implement & enforce modes for CA rotation
This gives the CA rotation action script two modes of which one must be
selected. In the first mode an additional issuer "next" is configured
and two CAs are considered valid by the k8s cluster components. Certs
get renewed by force and issued by the next issuer.
In the second mode, the CA rotation can be completed by dropping
basically invalidating the previous CA and allowing only the new one to
be considered valid. All certs and kubeconfigs will get renewed again as
we must restart all components anyway to make changes persistent (so
that they do not trust the previous CA anymore). We have to enforce the
renewal as the issued certs may still be in the valid timerange and
therefore would not get renewed which would skip restarting the
components.