Force renew K8s CP certificates during CA rotation
As one of the first steps we must renew all control plane component
certificates as the components get restarted in subsequent tasks.
The new certificates are issued by the next issuer if we started a
rotation. If we complete a rotation, they will be issued by the same
issuer which is then already called default :)