Renew calico certificates (legacy-setup)
Right after exchanging certs and CAs for Kubernetes itself, we must
update the calico stuff (legacy manifest-installation). Otherwise, the
CNI is broken and no new workload can be scheduled which would be bad as
we want to rollout restart nearly everything in a few moments.
The calico certs will also use the next issuer.