Properly generate CA certificate bundle
Don't obfuscate by setting an additional variable "ca_rotation".
Rework and simplify how we gather the CA certificates which allows us
to pull these via the approle and making a vault root token unnecessary
(the previous tasks required a root token).
Instead of using a boolean variable (ca_rotation) we pass the
(additional) issuer (if to be used) as "extra_ca". This simplifies the
code logic.