Properly generate issuers
We must add a new issuer to each PKI. As that's basically the same
functionality as when creating a cluster with root CAs, this his been
put into an own function.
Unfortunately, we can't supply an empty string as "issuer_name" (in the
hope that then the default will be used). Therefore, the if-else
branches have been created to either use the default or "next" issuer.