Fix kubeconfig server url
Referencing the frontend port via hostvars is actually a workaround
from ancient times when everything was a bit more messy. The problem
here is, that with restructuring the k8s-master role, that variable
actually can't be used. It is evaluated when the variable is referenced
but that fails ('available_frontend' is undefined). As we use an in-line
default when creating the kubeconfig (local_ipv4 + 6443), the default is
always used. However, we want the kubeconfigs to talk to the load balanced
K8s API and not to just their apiserver. There is no need anymore to do
the hostvars detour, the port is directly available as variable called
"k8s_apiserver_frontend_port".