Make the migration between local and http tf backend safe
(1) We ensure the uniqueness of the tf-state to guarantee
one source of truth by removing the out-to-date states
when switching between `local` and `http` Terraform backend methods.

(2) We do not run `terraform init` with `-migrate-state`
if a Terraform backend file exists on GitLab and `gitlab_backend=true`
to avoid overwriting the backend file with possibly outdated local data.

(3) We deleted default values from GitLab variables and now allow only
all GitLab variables are set OR all variables are unset
to clearly identify the user's intention.