Fix vrrp_priorities
Previously, the default values for vrrp_priorities where managed by a
role "vrrp-common". As that's role only purpose was to define values
for that variable, it has been moved to a vars/-file during the
core-split. However, as play-vars have higher precedence than inventory
variables, these values could not get overwritten via the config.toml
anymore. This commit fixes that by moving the default values to the
respective roles. I decided against a common-defaults role as previously
with the "vrrp-common" role, because the variable "vrrp_priorities" is
used in two roles only: keepalived and ch-k8s-lbaas-agent whereas one
resides in the k8s-core and the other in the k8s-supplements.