Add constrain to kubernetes-validate
The freshly released version v1.29.0 [1] is too recent and leads to
failures of the Ansible k8s module [2] triggered on Kubernetes nodes.
The jsonschema package fails with:
AttributeError: module 'jsonschema.validators' has no attribute 'Draft202012Validator'

Therefore, we're restricting the version of the kubernetes-validate
package to be the latest of the currently configured Kubernetes version
which is probably reasonable.

[1] https://pypi.org/project/kubernetes-validate/1.29.0/
[2] https://docs.ansible.com/ansible/latest/collections/kubernetes/core/k8s_module.html