Remove unused k8s-pki issuer role
I'm not sure what this role was intended to be used for; it is
referenced in the policy in init.sh for the orchestrator role, however,
the privileges granted by the certificate from this issue role are
useless. You can use it to impersonate a node, but that doesn't seem
like a useful thing to be able to do, in particular because the
orchestrator can create approle roles for nodes and then use that (which
is needed in the bare metal flow) to obtain credentials for the node.

Thus, we drop it, as well as the privileges to use it.