Don't set resource limits for NGINX ingress by default
Resource usage of the ingress controller can heavily vary between
clusters depending on the use case and workload. It is hard to come up
with limits for it therefore. Much worse than not setting limits and
getting DOSed which could potentially bring down the worker is from our
point of view setting insufficient limits which would have direct impact
on the availability of workload under legitimate normal circumstances.