Vaultify Thanos object storage configuration
As the Thanos object storage configuration contains confidential
information and should therefore not be plainly stored in a cluster
repository, it is now moved to Vault. This saves us from having to make
sure it is git-crypted. A migration function has been implemented to
easily move existing configurations to Vault. The documentation has been
extended to supply necessary information.
This is marked as a breaking change as an update of vault policies must
be executed for existing clusters.