Simplify Vault Kubernetes credential generation flow
For now, we don't talk about LDAP in particular and remove the
restriction on the common name of the obtained certificates. This allows
users to generate long-lived generic identities as well as short-lived
specific identities or a mixture of both.

Future work may introduce endpoints which have lower privileges than the
orchestrator role who are however able to obtain certificates tied to
their identity. Or, indeed, this is something people can do today
already.