Properly clean up Terraform states on destroy
The local terraform state files must not be removed. These get removed
automatically by Terraform.
Also the removal of the Gitlab state is moved to the end so that
a network hiccup does not block subsequent actions in which case one
would have to manually clean up resources in the OpenStack project.
We're also checking now the return code of our Gitlab state deletion.