Don't enforce setting/unsetting TF_HTTP_PASSWORD
TF_HTTP_PASSWORD is a user-specific environment variable in contrast to
the other variables which are checked here which are cluster-specific
configurations. As TF_HTTP_PASSWORD is user-specific and not cluster-
specific, it may be set in ~/.config/yaook-k8s/env (e.g. if the
application token is used for several clusters). We therefore cannot
simply enforce unsetting this variable. This would make the handling of
multiple clusters with differing backends unnecessarily complicated for
users.
However, a check has been added to check for this variable
whenever we try to pull the Gitlab state even if gitlab_backend=false.
In the other case (gitlab_backend=true), there is already a check in the
inventory updater.