CI: Create custom play to shuffle workload
In the CI we do only have two workers. In each run, these get drained
in serial which can fully load the first worker. That's a problem, as
we also define the first worker as test worker [1] and if it is already
full, new workload required for the tests can't be scheduled.
To try to balance workload (and to properly test the custom stage),
we're rollout restarting everything in the custom stage.